refactor(mcp): migrate handlers to lib/errors and lib/validation
