+++ plugins/maestro/src/mcp/handlers/session-state-tools.js
@@ -9,7 +9,8 @@ const {
   writeState,
   resolveStateDirPath,
 } = require('../../state/session-state');
-const { validateSessionId } = require('../../state/session-id-validator');
+const { assertSessionId } = require('../../lib/validation');
+const { ValidationError, StateError, NotFoundError } = require('../../lib/errors');
 
 const ACTIVE_SESSION_REL = path.join('state', 'active-session.md');
 
@@ -30,9 +31,7 @@ function getBasePath(projectRoot) {
 }
 
 function handleCreateSession(params, projectRoot) {
-  if (!validateSessionId(params.session_id)) {
-    throw new Error('Invalid session_id: must match pattern [a-zA-Z0-9_-]+');
-  }
+  assertSessionId(params.session_id);
 
   const basePath = getBasePath(projectRoot);
   const sessionPath = path.join(basePath, ACTIVE_SESSION_REL);
@@ -40,7 +39,7 @@ function handleCreateSession(params, projectRoot) {
   if (fs.existsSync(sessionPath)) {
     const existing = parseSessionState(fs.readFileSync(sessionPath, 'utf8'));
     if (existing.status === 'in_progress') {
-      throw new Error(
+      throw new StateError(
         `Active session '${existing.session_id}' already exists (status: in_progress). Archive it first with archive_session.`
       );
     }
@@ -152,8 +151,8 @@ function handleGetSessionStatus(_params, projectRoot) {
 }
 
 function handleTransitionPhase(params, projectRoot) {
-  if (params.session_id && !validateSessionId(params.session_id)) {
-    throw new Error('Invalid session_id: must match pattern [a-zA-Z0-9_-]+');
+  if (params.session_id) {
+    assertSessionId(params.session_id);
   }
 
   const basePath = getBasePath(projectRoot);
@@ -161,7 +160,7 @@ function handleTransitionPhase(params, projectRoot) {
   const state = parseSessionState(content);
 
   if (params.session_id && state.session_id !== params.session_id) {
-    throw new Error(
+    throw new StateError(
       `Session mismatch: active session is '${state.session_id}', got '${params.session_id}'`
     );
   }
@@ -170,14 +169,14 @@ function handleTransitionPhase(params, projectRoot) {
   const hasNextPhaseIds =
     Array.isArray(params.next_phase_ids) && params.next_phase_ids.length > 0;
   if (hasNextPhaseId && hasNextPhaseIds) {
-    throw new Error(
+    throw new ValidationError(
       'next_phase_id and next_phase_ids are mutually exclusive'
     );
   }
 
   const hasCompletedPhase = params.completed_phase_id != null;
   if (!hasCompletedPhase && !hasNextPhaseIds && !hasNextPhaseId) {
-    throw new Error(
+    throw new ValidationError(
       'At least one of completed_phase_id, next_phase_id, or next_phase_ids is required'
     );
   }
@@ -188,7 +187,7 @@ function handleTransitionPhase(params, projectRoot) {
       (phase) => phase.id === params.completed_phase_id
     );
     if (!completedPhase) {
-      throw new Error(
+      throw new NotFoundError(
         `Phase ${params.completed_phase_id} not found in session state`
       );
     }
@@ -198,7 +197,7 @@ function handleTransitionPhase(params, projectRoot) {
   if (hasNextPhaseId) {
     nextPhase = state.phases.find((phase) => phase.id === params.next_phase_id);
     if (!nextPhase) {
-      throw new Error(
+      throw new NotFoundError(
         `next_phase_id ${params.next_phase_id} does not match any phase in session state`
       );
     }
@@ -209,10 +208,10 @@ function handleTransitionPhase(params, projectRoot) {
     for (const id of params.next_phase_ids) {
       const phase = state.phases.find((candidate) => candidate.id === id);
       if (!phase) {
-        throw new Error(`Phase ${id} not found in session state`);
+        throw new NotFoundError(`Phase ${id} not found in session state`);
       }
       if (phase.status === 'completed' || phase.status === 'failed') {
-        throw new Error(`Cannot start phase ${id}: status is '${phase.status}'`);
+        throw new StateError(`Cannot start phase ${id}: status is '${phase.status}'`);
       }
       if (phase.status === 'pending') {
         phasesToStart.push(phase);
@@ -289,9 +288,7 @@ function handleTransitionPhase(params, projectRoot) {
 }
 
 function handleArchiveSession(params, projectRoot) {
-  if (!validateSessionId(params.session_id)) {
-    throw new Error('Invalid session_id: must match pattern [a-zA-Z0-9_-]+');
-  }
+  assertSessionId(params.session_id);
 
   const basePath = getBasePath(projectRoot);
   const sourcePath = path.join(basePath, ACTIVE_SESSION_REL);
@@ -299,7 +296,7 @@ function handleArchiveSession(params, projectRoot) {
   const state = parseSessionState(content);
 
   if (state.session_id !== params.session_id) {
-    throw new Error(
+    throw new StateError(
       `Session mismatch: active session is '${state.session_id}', requested '${params.session_id}'`
     );
   }
@@ -359,16 +356,14 @@ function handleArchiveSession(params, projectRoot) {
 }
 
 function handleUpdateSession(params, projectRoot) {
-  if (!validateSessionId(params.session_id)) {
-    throw new Error('Invalid session_id: must match pattern [a-zA-Z0-9_-]+');
-  }
+  assertSessionId(params.session_id);
 
   const basePath = getBasePath(projectRoot);
   const content = readState(ACTIVE_SESSION_REL, basePath);
   const state = parseSessionState(content);
 
   if (state.session_id !== params.session_id) {
-    throw new Error(
+    throw new StateError(
       `Session mismatch: active session is '${state.session_id}', got '${params.session_id}'`
     );
   }
@@ -388,7 +383,7 @@ function handleUpdateSession(params, projectRoot) {
   }
 
   if (updatedFields.length === 0) {
-    throw new Error(
+    throw new ValidationError(
       'At least one updatable field (execution_mode, execution_backend, current_batch) is required'
     );
   }
